UCS: better fit C11 standard
- removed void* arithmetics (replaced by macro)
- minor code cleaning
- this is part of PGI compiler adaptation

+++ src/ucs/sys/sys.c
@@ -526,7 +526,7 @@ size_t ucs_get_memfree_size()
     size_t mem_free;
 
     mem_free = ucs_get_meminfo_entry("MemFree");
-    if (mem_free == -1) {
+    if ((ssize_t)mem_free == -1) {
         mem_free = UCS_DEFAULT_MEM_FREE;
         ucs_info("cannot determine free mem size, using default: %zu",
                   mem_free);
